requirements.txt
# --- Web Framework & Server ---
fastapi==0.115.2
uvicorn[standard]==0.34.0

# --- HTTP Client (Async) ---
httpx==0.28.0

# --- Configuration & Validation ---
pydantic==2.7.4
pydantic-settings==2.12.0
PyYAML==6.0.2
python-dotenv==1.0.0

# --- Observability / Metrics ---
prometheus-client==0.21.0
prometheus-fastapi-instrumentator==7.0.0

# --- Structured Logging ---
structlog==24.2.0

# --- GPU / System Metrics ---
pynvml==11.5.0

# --- Background Jobs (optional, for scheduling health probes) ---
APScheduler==3.10.4

# --- Testing ---
pytest==8.3.2
pytest-asyncio==0.24.0

